requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
python-dotenv==1.0.0
schedule==1.2.0
lxml==4.9.3
pandas==2.1.4
webdriver-manager==4.0.1
twilio==8.10.0
flask==3.0.0
flask-cors==4.0.0
smtplib-ssl==1.0.2
